Refactored repeated circuit/abstract code into single files
