+++ plato.cabal
@@ -30,10 +30,8 @@ library
                       Tuura.Plato.Translation
     other-modules:    Tuura.Concept.Abstract,
                       Tuura.Concept.Circuit,
-                      Tuura.Concept.STG.Abstract,
-                      Tuura.Concept.STG.Circuit,
-                      Tuura.Concept.FSM.Abstract,
-                      Tuura.Concept.FSM.Circuit,
+                      Tuura.Concept.Circuit.Circuit,
+                      Tuura.Concept.Circuit.Abstract,
                       Tuura.PetriNet
     build-depends:    base >= 4.8 && < 5,
                       bifunctors <5.4 || >5.4,
