+++ src/Tuura/Concept/Circuit/Abstract.hs
@@ -1,4 +1,4 @@
-module Tuura.Concept.FSM.Abstract (
+module Tuura.Concept.Circuit.Abstract (
     Interface (..),
     InitialValue (..),
     Concept (..),
