+++ src/Tuura/Concept/STG.hs
@@ -1,9 +1,9 @@
 module Tuura.Concept.STG (
     module Data.Monoid,
-    module Tuura.Concept.STG.Abstract,
-    module Tuura.Concept.STG.Circuit,
+    module Tuura.Concept.Circuit.Abstract,
+    module Tuura.Concept.Circuit.Circuit,
     ) where
 
 import Data.Monoid
-import Tuura.Concept.STG.Abstract
-import Tuura.Concept.STG.Circuit
+import Tuura.Concept.Circuit.Abstract
+import Tuura.Concept.Circuit.Circuit
