+++ src/Tuura/Concept/STG/Simulation.hs
@@ -3,7 +3,7 @@ module Tuura.Concept.STG.Simulation (
     Simulation, PureSimulation, runSimulationSTG, runPureSimulationSTG
     ) where
 
-import Tuura.Concept.STG.Circuit
+import Tuura.Concept.Circuit.Circuit
 import Control.Monad.State hiding (State)
 import Data.Functor.Identity
